add a bunch of "f"s.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@19672 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/game/ShotManager.cxx
@@ -72,7 +72,7 @@ ShotManager::Shot::Shot(FiringInfo* info, int /*GUID*/, int /*p*/)
   lifetime = 0;
   range = 0;
 
-  for (size_t i = 0; i < 3; i++)
+  for (int i = 0; i < 3; i++)
     pos[i] = vec[i] = 0;
 }
